Consolite canvas helpers
