--- interface/CanvasStyle.h
@@ -1,22 +1,0 @@
-//
-//   @file    CanvasStyle.h         
-//   
-//            ATLAS Style, based on a style file from BaBar
-//
-//
-//   @author M.Sutton
-// 
-//   Copyright (C) 2010 Atlas Collaboration
-//
-//   $Id: CanvasStyle.h 289371 2012-05-23 16:24:19Z spagan $
-
-#ifndef  CanvasStyle_h
-#define CanvasStyle_h
-
-#include "TStyle.h"
-
-void SetCanvasStyle();
-
-TStyle* CanvasStyle(); 
-
-#endif // CanvasStyle_h
